requirements.txt
python-telegram-bot[job-queue]==21.5
openai==1.57.0
requests==2.32.3
psycopg2-binary==2.9.9
flask==3.0.0
